+++ lib/language_pack/rails3.rb
@@ -40,8 +40,18 @@ class LanguagePack::Rails3 < LanguagePack::Rails2
 
 private
 
-  def plugins
-    super.concat(%w( rails3_serve_static_assets )).uniq
+  def install_plugins
+    instrument "rails3.install_plugins" do
+      return false if gem_is_bundled?('rails_12factor')
+      plugins = {"rails_log_stdout" => "rails_stdout_logging", "rails3_serve_static_assets" => "rails_serve_static_assets" }.
+                 reject { |plugin, gem| gem_is_bundled?(gem) }
+      return false if plugins.empty?
+      plugins.each do |plugin, gem|
+        warn "Injecting plugin '#{plugin}'"
+      end
+      warn "Add 'rails_12factor' gem to your Gemfile to skip plugin injection"
+      LanguagePack::Helpers::PluginsInstaller.new(plugins.keys).install
+    end
   end
 
   # runs the tasks for the Rails 3.1 asset pipeline
@@ -67,8 +77,12 @@ private
               puts "Asset precompilation completed (#{"%.2f" % time}s)"
             else
               log "assets_precompile", :status => "failure"
+              deprecate <<-DEPRECATION
+                Runtime asset compilation is being removed on Sep. 18, 2013.
+                Builds will soon fail if assets fail to compile.
+              DEPRECATION
               puts "Precompiling assets failed, enabling runtime asset compilation"
-              install_plugin("rails31_enable_runtime_asset_compilation")
+              LanguagePack::Helpers::PluginsInstaller.new(["rails31_enable_runtime_asset_compilation"]).install
               puts "Please see this article for troubleshooting help:"
               puts "http://devcenter.heroku.com/articles/rails31_heroku_cedar#troubleshooting"
             end