+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v78 (8/28/2013)
+
+Features:
+
+* Don't add plugins if already gems
+
+Bugfixes:
+
+* Fix issue #127 Race condition with LPXC
+
 ## v77 (8/5/2013)
 
 Features: